mcp_servers/mpesa/pyproject.toml
[project]
name = "mpesa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.0",
    "python-dotenv>=1.1.1",
]
